# imobiliary-api/openapi.yaml
openapi: 3.1.0
info:
  title: Imobiliary API
  version: "0.9.0"
  summary: Rental property management for brokers, small agencies and owners.
  description: |
    The API behind the Imobiliary platform. It covers identity and the office
    (accounts, sessions with a second factor, members and invitations, the
    account holder's rights under the LGPD) and the office's register of
    people and properties. Contracts and rents arrive in later phases.

    ## Offices and roles

    Data belongs to an office (an organisation). An account works inside one
    office at a time and may belong to several. Inside an office it is either
    `admin`, who manages members, invitations and the office itself, or
    `member`. An office is never left without an administrator.

    ## Authentication

    A session is two credentials.

    - The **access token** is a JWT signed with Ed25519, valid for 15 minutes,
      sent as `Authorization: Bearer <token>`. It names the account and the
      office. The membership behind it is read again on every request, so a
      role change or a removal takes effect at once, not when the token
      expires. A token minted before the account's last password change is
      refused.
    - The **refresh token** is an opaque secret exchanged at
      `POST /v1/sessions/refresh` for a new pair. Each one is spent once.
      Presenting a spent one ends every session of the account, because two
...

    ## The second factor

...

...

    ## Errors

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearer: []
paths:
...
  /v1/rents/{rentID}/payment/preview:
    parameters:
      - name: rentID
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ID" }
    post:
      tags: [Rents]
      summary: Compute the late fee for a payment day
      description: |
        The penalty rate on the rent with its charges, plus the monthly
        interest rate prorated over a thirty-day month for each day after the
        due day, each rounded to the centavo. Paying on or before the due day
        adds nothing. Records nothing.
      operationId: previewPayment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paid_on]
              properties:
                paid_on: { $ref: "#/components/schemas/Date" }
      responses:
        "200":
          description: The late fee and the total.
          content:
            application/json:
              schema:
                type: object
                required: [late_fee, total]
                properties:
                  late_fee: { $ref: "#/components/schemas/LateFee" }
                  total: { $ref: "#/components/schemas/Money" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
  /v1/rents/{rentID}/charges:
    parameters:
      - name: rentID
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ID" }
    post:
      tags: [Rents]
      summary: Bill a charge with an unpaid instalment
      description: |
        Condominium, IPTU, water, energy or another charge, which needs a
        description. `destination` says where the money goes: `owner` enters
        the payout and pays the administration fee; `third_party` is paid on by
        the office (the condominium, the utility) and does neither. At most 20
        per instalment. Refused on `charges` once the instalment is paid. Once
        any instalment of a contract has a charge, replacing the contract's
        terms is refused on `rents`.
      operationId: addCharge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, amount, destination]
              properties:
                kind: { type: string, enum: [condominium, property_tax, water, energy, other] }
                description: { type: string, maxLength: 120 }
                amount: { $ref: "#/components/schemas/Money" }
                destination: { $ref: "#/components/schemas/ChargeDestination" }
      responses:
        "201":
          description: Added; the instalment as it is now.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RentDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /v1/rents/{rentID}/charges/{chargeID}:
    parameters:
      - name: rentID
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ID" }
      - name: chargeID
        in: path
        required: true
        schema: { $ref: "#/components/schemas/ID" }
    patch:
      tags: [Rents]
      summary: Change where a charge goes
      description: |
        The one change a charge still allows once the instalment is paid: its
        ledger lines are written again with the new destination, which is how
        charges recorded before the ledger are put right. Refused on
        `destination` once the instalment is in a payout. Audited.
      operationId: setChargeDestination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [destination]
              properties:
                destination: { $ref: "#/components/schemas/ChargeDestination" }
      responses:
        "200":
          description: Changed; the instalment as it is now.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RentDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "415": { $ref: "#/components/responses/UnsupportedMediaType" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
    delete:
      tags: [Rents]
      summary: Remove a charge from an unpaid instalment
      operationId: removeCharge
      responses:
        "200":
          description: Removed; the instalment as it is now.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RentDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
...
components:
...
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT (EdDSA)
  schemas:
    ID:
      type: string
      format: uuid
      examples: ["01923f6e-8b1a-7c3d-9e2f-4a5b6c7d8e9f"]
...
    PropertyAddress:
      type: object
      required: [street, city, state, zip_code]
      properties:
        street: { type: string, minLength: 1, maxLength: 120 }
        number: { type: string, maxLength: 20 }
        complement: { type: string, maxLength: 120 }
        district: { type: string, maxLength: 120 }
        city: { type: string, minLength: 1, maxLength: 120 }
        state: { type: string, description: "One of the 27 UFs, answered in capitals." }
        zip_code: { type: string, description: "CEP, answered as eight digits." }
        observation: { type: string, maxLength: 500 }
...
    Date:
      type: string
      format: date
      description: A calendar day, with no time zone.
      examples: ["2026-10-01"]
    Money:
      type: string
      description: Reais with two decimal places, never a number.
      pattern: '^[0-9]+\.[0-9]{2}$'
      examples: ["1500.00"]
...
    LateFee:
      type: object
      required: [days_late, penalty, interest, total]
      properties:
        days_late: { type: integer, minimum: 0 }
        penalty: { $ref: "#/components/schemas/Money" }
        interest: { $ref: "#/components/schemas/Money" }
        total: { $ref: "#/components/schemas/Money" }
    RentSummary:
      type: object
      required: [id, contract, sequence, due_on, amount, charges_total, due, late_fee, amount_paid, paid_on, income_tax_withheld, status]
      properties:
        id: { $ref: "#/components/schemas/ID" }
        contract:
          type: object
          required: [id, registry, address, tenant_names]
          properties:
            id: { $ref: "#/components/schemas/ID" }
            registry: { type: string }
            address: { $ref: "#/components/schemas/PropertyAddress" }
            tenant_names: { type: array, items: { type: string } }
        sequence: { type: integer, minimum: 1 }
        due_on: { $ref: "#/components/schemas/Date" }
        amount: { $ref: "#/components/schemas/Money", description: The rent of the instalment. }
        charges_total: { $ref: "#/components/schemas/Money" }
        due: { $ref: "#/components/schemas/Money", description: The rent with its charges. }
        late_fee: { $ref: "#/components/schemas/Money", description: As recorded with the payment; zero while unpaid. }
        amount_paid: { oneOf: [{ $ref: "#/components/schemas/Money" }, { type: "null" }] }
        paid_on: { oneOf: [{ $ref: "#/components/schemas/Date" }, { type: "null" }] }
        income_tax_withheld: { $ref: "#/components/schemas/Money", description: What a company tenant withheld; zero otherwise. }
        status: { type: string, enum: [paid, overdue, pending] }
    RentDetail:
      allOf:
        - $ref: "#/components/schemas/RentSummary"
        - type: object
          required: [charges, suggested_late_fee]
          properties:
            charges:
              type: array
              items:
                type: object
                required: [id, kind, description, amount, destination]
                properties:
                  id: { $ref: "#/components/schemas/ID" }
                  kind: { type: string, enum: [condominium, property_tax, water, energy, other] }
                  description: { type: string }
                  amount: { $ref: "#/components/schemas/Money" }
                  destination: { $ref: "#/components/schemas/ChargeDestination" }
            suggested_late_fee: { $ref: "#/components/schemas/LateFee" }
    ChargeDestination:
      type: string
      enum: [owner, third_party]
      description: Where a charge's money goes once paid.
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - validation_failed
                - not_found
                - conflict
                - unauthorized
                - forbidden
                - mfa_required
                - mfa_enrollment_required
                - rate_limited
                - bad_request
                - unsupported_media_type
                - payload_too_large
                - internal_error
                - unavailable
                - precondition_failed
                - precondition_required
                - in_use
            message: { type: string }
            fields:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
  responses:
    BadRequest:
      description: The body could not be read, or is larger than 1 MiB (`413`).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: |
        Authentication failed. One answer for a missing or refused token, a
        wrong password or code, and a spent or unknown link.
      headers:
        WWW-Authenticate:
          description: Sent when the bearer token was missing or refused.
          schema: { type: string, examples: ["Bearer"] }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: |
        `forbidden`: the caller may not do this, or does not belong to the
        office. `mfa_enrollment_required`: the session must enrol a second
        factor first.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource in this office.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    UnsupportedMediaType:
      description: The body must be `application/json`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationFailed:
      description: The request was read and refused. `fields` lists every problem.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    RateLimited:
      description: Too many requests.
      headers:
        Retry-After:
          description: Whole seconds to wait.
          schema: { type: integer, minimum: 1 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }